refactor: Move these tests
