+++ app/src/test/java/com/openblocks/blocks/view/example/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.openblocks.blocks.view;
+package com.openblocks.blocks.view.example;
 
 import org.junit.Test;
